Fock Operators

Fock Operators

The Fock operator consists of an AbstractQuantumSystem and a set of coupled integro-differential equations, the solution of which is the objective of the self-consistent field procedure. For the solution process to work, the set of equations need to implement a few methods:

It must fulfil Julia's iteration interface, i.e. each element must be the equation for a single orbital, which is solved independently from the other equations, but with the other orbitals as inputs.

energy which calculates the orbital energy, i.e. $\matrixel{\chi}{\Hamiltonian}{\chi}$.

energy_matrix! which calculates the energy matrix $\mat{H}_i$ for orbital equation $i$, where $\vec{c}^H\mat{H}_i\vec{c}$ gives the orbital energy for the correspond orbital. The overall energy matrix $\mat{H}=\sum_i\mat{H}_i$ is used to solve the secular problem for the mixing coefficients.

update! which recomputes all orbital-dependent integrals, shared among the equations of the equation system.

SCF.Fock — Type.

Fock(quantum_system, equations)

A Fock operator consists of a quantum_system, from which equations are variationally derived. equations must be iterable, where each element corresponds to the equation for one orbital, and must provide overloads for the functions energy, energy_matrix!, and hamiltonian. Additionally, update! must be provided for equations, to prepare the equation system for the next iteration.

source

SCF.norm_rot! — Function.

norm_rot!(fock, v)

Normalize and rotate the eigenvector v such that the largest lobe has positive sign.

source

SCF.rotate_max_lobe! — Function.

rotate_max_lobe!(v)

Rotate the vector v such that the largest lobe has positive sign.

source

SCF.energy — Function.

energy(equation::Equation)

Computes the orbital energy of equation. To be overloaded by the user.

source

energy(fock::Fock)

Calculates the total energy of the system by summing the orbital energies.

source

SCF.energy_matrix! — Function.

energy_matrix!(H::AbstractMatrix, equation::Equation)

Assemble the energy matrix for equation, i.e. the matrix where the elements are the energies corresponding to the mixing of configurations, basically the Hamiltonian matrix of the corresponding orbital. This matrix, sandwiched between the mixing coefficients, gives the orbital energy. To be overloaded by the user. NB the matrix elements should be added to, i.e. not overwritten, rather, the overload of energy_matrix! should compute (the equivalent of) H += H_i, where H_i is the energy matrix for equation.

source

energy_matrix!(H::AbstractMatrix, fock::Fock)

Calculates the total energy matrix of the system by summing the energy matrices of the different orbital equations of fock. This overwrites the entries of H.

source

SCF.hamiltonian — Function.

hamiltonian(equation::Equation)

Returns the orbital Hamiltonian of equation. To be overloaded by the user.

source

SCF.update! — Function.

update!(eqs; kwargs...)

Update the equation system eqs for the current iteration. To be overloaded by the user.

source