Fock Operators

Fock Operators

SCF.Fock — Type.

Fock(quantum_system, equations)

A Fock operator consists of a quantum_system, from which equations are variationally derived.

source

SCF.norm_rot! — Function.

norm_rot!(fock, v)

Normalize and rotate the eigenvector v such that the largest lobe has positive sign.

source

SCF.rotate_max_lobe! — Function.

rotate_max_lobe!(v)

Rotate the vector v such that the largest lobe has positive sign.

source