Self-Consistent Iteration

Self-Consistent Iteration

SCF.scf! — Function.

scf!([fun!, ]fock[; ω=0, max_iter=200, tol=1e-8, verbosity=1])

Perform the SCF iterations for the fock operator. One iteration is performed by scf_iteration!. The optional fun!(P̃,c̃) argument allows for extra operations to be performed every SCF cycle (such as plotting, etc).

ω is a relaxation parameter; the orbitals and mixing coefficients are updated as wᵢ₊₁ = (1-ω)w̃ + ωwᵢ where w̃ is the solution in the current iteration and wᵢ the previous solution. The default (ω=0) is to only use this solution. A larger value of ω makes it easier to achieve convergence in oscillatory problems, but a too large value decreases the convergence rate. It is therefore desireable to reduce the value of ω if the convergence is deemed to be monotonous, the criterion for which is whether the change over the monotonous_window last iterations is steadily decreasing. If this is the case, ω will be reduced by multiplying by ωfactor. Conversely, if the change is non-monotonous, ω will be increased, but not beyond ωmax.

The SCF procedure continues until either the amount of iterations equals max_iter or the change in the coefficients is below tol.

source

SCF.scf_iteration! — Function.

scf_iteration!(fock, P, c[; verbosity=0, method=:arnoldi, σ=nothing, tol=1e-10,
               update_mixing_coefficients=true])

Perform one step of the SCF iteration. This will

Improve each of the orbitals in turn, using the values of the orbitals P and mixing coefficients c from the previous step as input.

Solve the secular problem to find new values for the mixing coefficients, c, unless update_mixing_coefficients==false.

source

SCF.solve_orbital_equation! — Function.

solve_orbital_equation!(Pj, eq, method, tol)

Solve the orbital equation eq and store the result in the radial orbital vector of expansion coefficients Pj. The solution is computed using method; valid choices are

:arnoldi, which requires that hamiltonian(eq) supports KrylovWrapper, and

:arnoldi_shift_invert, which iterates (H-σ*I)⁻¹ and requires that hamiltonian(eq) supports KrylovWrapper and provides an overload for MatrixFactorizations.preconditioner. The shift is automatically chosen as 1.1ϵ where ϵ is the (current estimate of the) orbital energy of the orbital governed by eq.

Both methods are controlled by the stopping tolerance tol.

source

SCF.solve_secular_problem! — Function.

solve_secular_problem!(H, c, fock)

Form the energy matrix, store it in H, and then solve the secular problem Hc = Ec for the lowest eigenvalue.

source

KrylovWrapper

The orbital improvements are performed via diagonalization of the integro-differential equations for each orbital. The diagonalization procedure is the Arnoldi method, which requires the repeated action of the Hamiltonian on a vector. The KrylovWrapper object is used to wrap Hamiltonians which are not simple <:AbstractMatrix objects, but which require a specialized implementation of mul! to act on a <:AbstractVector of coefficients. An example is the kind of Hamiltonians implemented in the Atoms.jl library, which are based on the function space algebra from ContinuumArrays.jl.

SCF.KrylovWrapper — Type.

KrylovWrapper(hamiltonian)

Proxy object used in the Krylov iterations, during orbital improvement. This is useful, since hamiltonian may be defined to act on objects such as vectors living in function spaces (as e.g. implemented using ContinuumArrays.jl), whereas the SCF iterations act on the coefficients directly.

source

Base.size — Method.

size(::KrylovWrapper)

Returns the dimension of the KrylovWrapper. For Hamiltonians which are not <:AbstractMatrix, this needs to be overloaded.

source

LinearAlgebra.mul! — Method.

mul!(y, ::KrylovWrapper, x)

Compute the action of the wrapped Hamiltonian on x and store it in y. For Hamiltonians which are not <:AbstractMatrix, this needs to be overloaded.

source

Shift-and-invert

To improve the rate of convergence of the Krylov iterations, instead of computing the eigenvectors of $\Hamiltonian$ when improving the orbitals, we can use the shift-and-invert strategy and instead iterate $(\Hamiltonian-\sigma I)^{-1}$ with a shift $\sigma$ that is located slightly below the target eigenvalue. The eigenvectors are shared between $\Hamiltonian$ and $(\Hamiltonian-\sigma I)^{-1}$, but convergence towards the low-lying states is faster for $(\Hamiltonian-\sigma I)^{-1}$.

This approach is complicated by the fact that the orbital Hamiltonian $\Hamiltonian$ is an integro-differential operator, which furthermore depends on the other orbitals through direct, exchange, and configuration interaction, and thus $\Hamiltonian$ is not easily factorizable. However, we can compute the action of $(\Hamiltonian-\sigma I)^{-1}$ using an iterative solver with a preconditioner constructed from all terms of $\Hamiltonian$ except the integral operators (exchange interaction) and the source terms (configuration interaction). As mentioned in the documentation for solve_orbital_equation!, the orbital Hamiltonian must support KrylovWrapper as well as providing an overload for MatrixFactorizations.preconditioner.

SCF.ShiftInvert — Type.

ShiftInvert(A⁻¹, σ)

Structure representing the inverse of a shifted matrix Ã = A + σ*I. This is used for finding eigenpairs of Ã in the neighbourhood of σ; useful in power methods such as the Krylov iteration, which converge the convex hull of the eigenspectrum first.

source

SCF.ShiftInvert — Method.

ShiftInvert(A[, σ])

Construct the ShiftInvert structure from the factorization of the matrix A shifted by σ. If no shift is specified, the lower bound as calculated by gershgorin_bounds is used by default.

source

LinearAlgebra.mul! — Method.

mul!(y, si::ShiftInvert, x)

Compute the action of the shifted-and-inverted matrix si.A⁻¹ on x and store the result in y. It is assumed that the factorization does not depend on the initial contents of y; if that is not the case, an overload has to be provided.

source

LinearAlgebra.mul! — Method.

mul!(y, si::ShiftInvert{<:IterativeFactorization}, x)

Compute the action of the shifted-and-inverted matrix si.A⁻¹ on x and store the result in y for factorizations based on IterativeSolvers.jl.

source

Gershgorin's circle theorem

Gershgorin's circle theorem can be used to estimate the bounds on the eigenspectrum of a matrix. It is always correct, but only a useful estimate when the matrix is diagonally dominant. Furthermore, gershgorin_bounds is limited to the case of Hermitian matrices, whose eigenvalues lie on the real line. If an experimental energy is known, that is almost always a better lower bound to use for the Shift-and-invert strategy.

SCF.gershgorin_disc — Function.

gershgorin_disc(i, row)

Return the Gershgorin disc for the ith row of a matrix.

source

SCF.gershgorin_discs — Function.

gershgorin_discs(A)

Compute all Gershgorin discs of a matrix A.

source

SCF.gershgorin_bounds — Function.

gershgorin_bounds(A)

For a Hermitian matrix A, compute the bounds on the real axis of the eigenspectrum. NB that this bound will only be useful for diagonally dominant As.

source