Self-Consistent Iteration

Self-Consistent Iteration

SCF.scf! — Function.

scf!([fun!, ]fock[; ω=0, max_iter=200, tol=1e-8, verbosity=1])

Perform the SCF iterations for the fock operator. One iteration is performed by scf_iteration!. The optional fun!(P̃,c̃) argument allows for extra operations to be performed every SCF cycle (such as plotting, etc).

ω is a relaxation parameter; the orbitals and mixing coefficients are updated as wᵢ₊₁ = (1-ω)w̃ + ωwᵢ where w̃ is the solution in the current iteration and wᵢ the previous solution. The default (ω=0) is to only use this solution.

The SCF procedure continues until either the amount of iterations equals max_iter or the change in the coefficients is below tol.

source

SCF.scf_iteration! — Function.

scf_iteration!(fock, P, c[; verbosity=0, method=:arnoldi, tol=1e-10,
               update_mixing_coefficients=true])

Perform one step of the SCF iteration. This will

Improve each of the orbitals in turn, using the values of the orbitals P and mixing coefficients c from the previous step as input.

Solve the secular problem to find new values for the mixing coefficients, c, unless update_mixing_coefficients==false.

source

KrylovWrapper

The orbital improvements are performed via diagonalization of the integro-differential equations for each orbital. The diagonalization procedure is the Arnoldi method, which requires the repeated action of the Hamiltonian on a vector. The KrylovWrapper object is used to wrap Hamiltonians which are not simple <:AbstractMatrix objects, but which require a specialized implementation of mul! to act on a <:AbstractVector of coefficients. An example is the kind of Hamiltonians implemented in the Atoms.jl library, which are based on the function space algebra from ContinuumArrays.jl.

SCF.KrylovWrapper — Type.

KrylovWrapper(hamiltonian)

Proxy object used in the Krylov iterations, during orbital improvement. This is useful, since hamiltonian may be defined to act on objects such as vectors living in function spaces (as e.g. implemented using ContinuumArrays.jl), whereas the SCF iterations act on the coefficients directly.

source

Base.size — Method.

size(::KrylovWrapper)

Returns the dimension of the KrylovWrapper. For Hamiltonians which are not <:AbstractMatrix, this needs to be overloaded.

source

LinearAlgebra.mul! — Function.

mul!(y, ::KrylovWrapper, x)

Compute the action of the wrapped Hamiltonian on x and store it in y. For Hamiltonians which are not <:AbstractMatrix, this needs to be overloaded.

source